Recovered from Eclipse.
git-svn-id: https://subversion.assembla.com/svn/cmosher@1966 884d43a4-ffbe-47e9-90ca-fd8c37d7eb32

+++ core50/src/nu/mine/mosher/core/Util.java
@@ -25,7 +25,7 @@ public final class Util
 
 	public static void unused(double i)
 	{
-		long x = i;
+		double x = i;
 		i = x;
 	}
 }
